feat(volo-http): add feature for switching json lib (#308)
* feat(volo-http): add feature for switching json lib

This commit adds two features `serde_json` and `sonic_json` to use
`serde_json` or `sonic-rs`.

Note that you cannot have more than one json libraries. If the
relevant features are not enabled, json will not be available.

* chore(volo-http): enable `sonic_json` by default

---------
