+++ Cargo.toml
@@ -96,6 +96,7 @@ serde_urlencoded = "0.7"
 serde_yaml = "0.9"
 simdutf8 = "0.1"
 socket2 = "0.5"
+sonic-rs = "0.3"
 syn = "1"
 tempfile = "3"
 thiserror = "1"
